+++ app/pages/5qcard/FiveQPage.tsx
@@ -26,7 +26,7 @@ import Front5Q from './5QFront';
 import Back5Q from './5QBack';
 import PageContainer from '../../globals/PageContainer';
 
-const FiveQpage: React.FunctionComponent = () => {
+const FiveQPage: React.FunctionComponent = () => {
     const [front, setFront] = useState<boolean>(true);
 
     return (
@@ -48,4 +48,4 @@ const FiveQpage: React.FunctionComponent = () => {
     );
 };
 
-export default FiveQpage;
+export default FiveQPage;