+++ CHANGELOG.md
@@ -25,6 +25,7 @@ This file is offered as-is, without any warranty.
 
 ## [Unreleased](https://github.com/marcoXbresciani/TKCompanionApp/commits/HEAD)
 ### Added
+* Support for dark theme.
 ### Changed
 * Default card layout also for 5Q card.
 ### Removed